refactor: torus auth
